Extract revoke/update functions out of useRevoke hook
